AI_End/requirements.txt
torch
torchvision
numpy
scikit-learn
pillow
joblib
fastapi
uvicorn